Rename Quartznet.preprocess to transform
