--- docs/api/Quartznet/preprocess.md
@@ -1,1 +1,0 @@
-::: thunder.quartznet.preprocess
